# backend/docs/openapi.json
{"openapi":"3.0.1","info":{"title":"ScaleVision Backend API","description":"Contrato técnico del backend MVP para procesamiento de video con IA","contact":{"name":"ScaleVision Team"},"version":"1.0.0"},"servers":[{"url":"http://localhost:8080","description":"Servidor de desarrollo local"}],"tags":[{"name":"Callbacks","description":"Endpoints para recibir callbacks de la IA con el resultado del procesamiento"},{"name":"Discovery","description":"Endpoints para escanear sujetos en un video antes del procesamiento"},{"name":"Job Status","description":"Endpoints para consultar el estado de un job de procesamiento"},{"name":"Video Processing","description":"Endpoints para iniciar el procesamiento de video con IA"},{"name":"AI Worker","description":"Endpoints para consultar estado y salud del worker de IA"}],"paths":{"/svmvp/videos/process":{"post":{"tags":["Video Processing"],"summary":"Iniciar procesamiento de video","description":"Envía un video a la IA para su procesamiento. Retorna un jobId para hacer polling del estado.","operationId":"processVideo","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProcessVideoRequest"}}},"required":true},"responses":{"400":{"description":"Request inválido","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProcessVideoResponse"}}}},"422":{"description":"Error al enviar video a la IA","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProcessVideoResponse"}}}},"202":{"description":"Job creado y en procesamiento","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProcessVideoResponse"}}}}}}},"/videos/process":{"post":{"tags":["Video Processing"],"summary":"Iniciar procesamiento de video","description":"Envía un video a la IA para su procesamiento. Retorna un jobId para hacer polling del estado.","operationId":"processVideo_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProcessVideoRequest"}}},"required":true},"responses":{"400":{"description":"Request inválido","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProcessVideoResponse"}}}},"422":{"description":"Error al enviar video a la IA","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProcessVideoResponse"}}}},"202":{"description":"Job creado y en procesamiento","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProcessVideoResponse"}}}}}}},"/callbacks/ai":{"post":{"tags":["Callbacks"],"summary":"Recibir callback de la IA","description":"La IA llama a este endpoint cuando termina el procesamiento. Requiere X-Webhook-Secret válido.","operationId":"handleAiCallback","parameters":[{"name":"X-AI-Schema-Version","in":"header","required":false,"schema":{"type":"string"}},{"name":"X-Webhook-Secret","in":"header","required":false,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AICallbackRequest"}}},"required":true},"responses":{"400":{"description":"Request inválido o schema version incorrecto","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CallbackResponse"}}}},"200":{"description":"Callback procesado correctamente","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CallbackResponse"}}}},"404":{"description":"Job no encontrado","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CallbackResponse"}}}},"401":{"description":"Webhook secret inválido","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CallbackResponse"}}}},"409":{"description":"Transición de estado inválida","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CallbackResponse"}}}}}}},"/svmvp/callbacks/ai":{"post":{"tags":["Callbacks"],"summary":"Recibir callback de la IA","description":"La IA llama a este endpoint cuando termina el procesamiento. Requiere X-Webhook-Secret válido.","operationId":"handleAiCallback_1","parameters":[{"name":"X-AI-Schema-Version","in":"header","required":false,"schema":{"type":"string"}},{"name":"X-Webhook-Secret","in":"header","required":false,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AICallbackRequest"}}},"required":true},"responses":{"400":{"description":"Request inválido o schema version incorrecto","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CallbackResponse"}}}},"200":{"description":"Callback procesado correctamente","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CallbackResponse"}}}},"404":{"description":"Job no encontrado","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CallbackResponse"}}}},"401":{"description":"Webhook secret inválido","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CallbackResponse"}}}},"409":{"description":"Transición de estado inválida","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CallbackResponse"}}}}}}},"/svmvp/ai/scan-subjects":{"post":{"tags":["Discovery"],"summary":"Escanear sujetos en video","description":"Analiza el video y retorna los sujetos detectados con su ubicación y tipo.","operationId":"scanSubjects","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ScanSubjectsRequest"}}},"required":true},"responses":{"400":{"description":"Request inválido","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ScanSubjectsResponse"}}}},"200":{"description":"Sujetos detectados correctamente","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ScanSubjectsResponse"}}}}}}},"/ai/scan-subjects":{"post":{"tags":["Discovery"],"summary":"Escanear sujetos en video","description":"Analiza el video y retorna los sujetos detectados con su ubicación y tipo.","operationId":"scanSubjects_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ScanSubjectsRequest"}}},"required":true},"responses":{"400":{"description":"Request inválido","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ScanSubjectsResponse"}}}},"200":{"description":"Sujetos detectados correctamente","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ScanSubjectsResponse"}}}}}}},"/jobs/{jobId}":{"get":{"tags":["Job Status"],"summary":"Consultar estado del job","description":"Retorna el estado actual del job, progreso y resultado si ya completó.","operationId":"getJobStatus","parameters":[{"name":"jobId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"Estado del job retornado correctamente","content":{"*/*":{"schema":{"$ref":"#/components/schemas/JobStatusResponse"}}}},"400":{"description":"jobId inválido","content":{"*/*":{"schema":{"$ref":"#/components/schemas/JobStatusResponse"}}}},"404":{"description":"Job no encontrado","content":{"*/*":{"schema":{"$ref":"#/components/schemas/JobStatusResponse"}}}}}}},"/svmvp/jobs/{jobId}":{"get":{"tags":["Job Status"],"summary":"Consultar estado del job","description":"Retorna el estado actual del job, progreso y resultado si ya completó.","operationId":"getJobStatus_1","parameters":[{"name":"jobId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"Estado del job retornado correctamente","content":{"*/*":{"schema":{"$ref":"#/components/schemas/JobStatusResponse"}}}},"400":{"description":"jobId inválido","content":{"*/*":{"schema":{"$ref":"#/components/schemas/JobStatusResponse"}}}},"404":{"description":"Job no encontrado","content":{"*/*":{"schema":{"$ref":"#/components/schemas/JobStatusResponse"}}}}}}},"/svmvp/ai/job/{jobId}":{"get":{"tags":["AI Worker"],"summary":"Consultar estado de job en IA","description":"Fallback para consultar el estado del job directamente en el worker de IA.","operationId":"getWorkerJobStatus","parameters":[{"name":"jobId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"404":{"description":"Job no encontrado en el worker","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AIWorkerJobStatusResponse"}}}},"200":{"description":"Estado retornado correctamente","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AIWorkerJobStatusResponse"}}}}}}},"/ai/job/{jobId}":{"get":{"tags":["AI Worker"],"summary":"Consultar estado de job en IA","description":"Fallback para consultar el estado del job directamente en el worker de IA.","operationId":"getWorkerJobStatus_1","parameters":[{"name":"jobId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"404":{"description":"Job no encontrado en el worker","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AIWorkerJobStatusResponse"}}}},"200":{"description":"Estado retornado correctamente","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AIWorkerJobStatusResponse"}}}}}}},"/svmvp/ai/health":{"get":{"tags":["AI Worker"],"summary":"Health check del worker de IA","description":"Verifica que el worker de IA esté disponible.","operationId":"getWorkerHealth","responses":{"200":{"description":"Worker disponible","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AIWorkerHealthResponse"}}}}}}},"/ai/health":{"get":{"tags":["AI Worker"],"summary":"Health check del worker de IA","description":"Verifica que el worker de IA esté disponible.","operationId":"getWorkerHealth_1","responses":{"200":{"description":"Worker disponible","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AIWorkerHealthResponse"}}}}}}}},"components":{"schemas":{"Config":{"type":"object","properties":{"targetAspectRatio":{"type":"string"},"fpsSampled":{"type":"integer","format":"int32"},"includeTrajectoryData":{"type":"boolean"}}},"ProcessVideoRequest":{"required":["videoUrl"],"type":"object","properties":{"videoUrl":{"pattern":"^https?://.*","type":"string"},"targetAspectRatio":{"type":"string"},"targetDuration":{"type":"integer","format":"int32"},"focusArea":{"type":"string"},"callbackUrl":{"pattern":"^https?://.*","type":"string"},"webhookSecret":{"type":"string"},"trackingMode":{"type":"string"},"targetSelection":{"$ref":"#/components/schemas/TargetSelection"},"config":{"$ref":"#/components/schemas/Config"}}},"TargetSelection":{"type":"object","properties":{"referenceTimestamp":{"type":"number","format":"double"},"referenceBox":{"type":"array","items":{"type":"number","format":"double"}}}},"ProcessVideoResponse":{"type":"object","properties":{"jobId":{"type":"string","format":"uuid"},"status":{"type":"string"}}},"AICallbackRequest":{"required":["jobId","status"],"type":"object","properties":{"jobId":{"type":"string"},"status":{"type":"string"},"outputUrl":{"type":"string"},"errorCode":{"type":"string"},"retryable":{"type":"boolean"},"processingStats":{"type":"object"},"cropRecommendations":{"type":"object"}}},"CallbackResponse":{"type":"object","properties":{"jobId":{"type":"string","format":"uuid"},"status":{"type":"string"}}},"ScanSubjectsRequest":{"required":["videoUrl"],"type":"object","properties":{"videoUrl":{"pattern":"^https?://.*","type":"string"},"minAppearanceRatio":{"type":"number","format":"double"},"config":{"$ref":"#/components/schemas/Config"}}},"ReferenceData":{"type":"object","properties":{"timestamp":{"type":"number","format":"double"},"box":{"type":"array","items":{"type":"number","format":"double"}}}},"ScanSubjectsResponse":{"type":"object","properties":{"status":{"type":"string"},"subjects":{"type":"array","items":{"$ref":"#/components/schemas/SubjectData"}}}},"SubjectData":{"type":"object","properties":{"tempId":{"type":"string"},"thumbnailBase64":{"type":"string"},"appearanceRatio":{"type":"number","format":"double"},"subjectType":{"type":"string"},"referenceData":{"$ref":"#/components/schemas/ReferenceData"}}},"ErrorData":{"type":"object","properties":{"code":{"type":"string"},"message":{"type":"string"},"retryable":{"type":"boolean"}}},"JobStatusResponse":{"type":"object","properties":{"jobId":{"type":"string","format":"uuid"},"status":{"type":"string"},"progressPercentage":{"type":"integer","format":"int32"},"result":{"$ref":"#/components/schemas/ResultData"},"error":{"$ref":"#/components/schemas/ErrorData"}}},"ResultData":{"type":"object","properties":{"outputVideoUrl":{"type":"string"},"aiResultPayload":{"type":"string"}}},"AIWorkerJobStatusResponse":{"type":"object","properties":{"jobId":{"type":"string"},"status":{"type":"string"},"progressPercentage":{"type":"integer","format":"int32"},"rawResult":{"type":"string"}}},"AIWorkerHealthResponse":{"type":"object","properties":{"status":{"type":"string"},"activeJobs":{"type":"integer","format":"int32"},"capacityAvailable":{"type":"boolean"}}}}}}