# backend/src/main/java/com/scalevision/backend/infrastructure/adapter/in/rest/CallbackController.java
package com.scalevision.backend.infrastructure.adapter.in.rest;

import com.scalevision.backend.application.port.out.JobRepository;
import com.scalevision.backend.domain.exception.InvalidStatusTransitionException;
import com.scalevision.backend.domain.model.JobStatus;
import com.scalevision.backend.domain.model.ProcessingJob;
import com.scalevision.backend.infrastructure.adapter.in.rest.dto.AICallbackRequest;
import com.scalevision.backend.infrastructure.adapter.in.rest.dto.CallbackResponse;
import com.fasterxml.jackson.databind.ObjectMapper;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.ExampleObject;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.responses.ApiResponses;
import io.swagger.v3.oas.annotations.tags.Tag;
import jakarta.validation.Valid;
import org.springframework.beans.factory.annotation.Value;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RestController;
import org.springframework.web.server.ResponseStatusException;

import java.util.UUID;

import static org.springframework.http.HttpStatus.BAD_REQUEST;
import static org.springframework.http.HttpStatus.CONFLICT;
import static org.springframework.http.HttpStatus.NOT_FOUND;

@RestController
@Tag(name = "Callbacks", description = "Endpoints para recibir callbacks de la IA con el resultado del procesamiento")
public class CallbackController {

    private final JobRepository jobRepository;
    private final String expectedSchemaVersion;
    private final ObjectMapper objectMapper;

    public CallbackController(
            JobRepository jobRepository,
            @Value("${ai.callback.schema-version:1.0.0}") String expectedSchemaVersion
    ) {
        this.jobRepository = jobRepository;
        this.expectedSchemaVersion = expectedSchemaVersion;
        this.objectMapper = new ObjectMapper();
    }

    @Operation(
            summary = "Recibir callback de la IA",
            description = "La IA llama a este endpoint cuando termina el procesamiento. Requiere X-Webhook-Secret válido."
    )
    @ApiResponses({
            @ApiResponse(responseCode = "200", description = "Callback procesado correctamente",
                    content = @Content(examples = @ExampleObject(value = """
                        {
                          "jobId": "550e8400-e29b-41d4-a716-446655440000",
                          "status": "COMPLETED"
                        }
                        """))),
            @ApiResponse(responseCode = "400", description = "Request inválido o schema version incorrecto",
                    content = @Content(examples = @ExampleObject(value = """
                        {
                          "code": "BAD_REQUEST",
                          "message": "schema version is invalid",
                          "timestamp": "2026-02-21T12:00:00"
                        }
                        """))),
            @ApiResponse(responseCode = "401", description = "Webhook secret inválido",
                    content = @Content(examples = @ExampleObject(value = """
                        {
                          "code": "HTTP_401",
                          "message": "invalid webhook secret",
                          "timestamp": "2026-02-21T12:00:00"
                        }
                        """))),
            @ApiResponse(responseCode = "404", description = "Job no encontrado",
                    content = @Content(examples = @ExampleObject(value = """
                        {
                          "code": "HTTP_404",
                          "message": "job not found",
                          "timestamp": "2026-02-21T12:00:00"
                        }
                        """))),
            @ApiResponse(responseCode = "409", description = "Transición de estado inválida",
                    content = @Content(examples = @ExampleObject(value = """
                        {
                          "code": "HTTP_409",
                          "message": "invalid status transition",
                          "timestamp": "2026-02-21T12:00:00"
                        }
                        """)))
    })
    @PostMapping({"/callbacks/ai", "/svmvp/callbacks/ai"})
    public ResponseEntity<CallbackResponse> handleAiCallback(
            @Valid @RequestBody AICallbackRequest request,
            @RequestHeader(name = "X-AI-Schema-Version", required = false) String schemaVersion,
            @RequestHeader(name = "X-Webhook-Secret", required = false) String webhookSecret
    ) {
        UUID jobId = parseJobId(request.jobId());

        ProcessingJob job = jobRepository.findById(jobId)
                .orElseThrow(() -> new ResponseStatusException(NOT_FOUND, "job not found"));

        if (job.getWebhookSecret() != null && !job.getWebhookSecret().isBlank()) {
            if (webhookSecret == null || !job.getWebhookSecret().equals(webhookSecret)) {
                throw new ResponseStatusException(org.springframework.http.HttpStatus.UNAUTHORIZED, "invalid webhook secret");
            }
        }

        if (schemaVersion != null && !expectedSchemaVersion.equals(schemaVersion)) {
            throw new ResponseStatusException(BAD_REQUEST, "schema version is invalid");
        }

        JobStatus nextStatus = mapStatus(request.status());

        try {
            job.updateStatus(nextStatus);
        } catch (InvalidStatusTransitionException ex) {
            throw new ResponseStatusException(CONFLICT, ex.getMessage());
        }

        if (nextStatus == JobStatus.COMPLETED) {
            job.setOutputUrl(request.outputUrl());
            job.setAiResultPayload(buildSuccessResultJson(request));
        }

        if (nextStatus == JobStatus.FAILED) {
            job.setErrorMessage(request.errorCode());
            job.setAiResultPayload(buildFailureResultJson(request));
        }

        jobRepository.save(job);

        return ResponseEntity.ok(new CallbackResponse(job.getId(), job.getStatus().name()));
    }

    private UUID parseJobId(String rawJobId) {
        try {
            return UUID.fromString(rawJobId);
        } catch (Exception ex) {
            throw new ResponseStatusException(BAD_REQUEST, "jobId is invalid");
        }
    }

    private JobStatus mapStatus(String rawStatus) {
        String normalized = rawStatus.trim().toUpperCase();

        if ("COMPLETED".equals(normalized)) {
            return JobStatus.COMPLETED;
        }

        if ("FAILED".equals(normalized)) {
            return JobStatus.FAILED;
        }

        throw new ResponseStatusException(BAD_REQUEST, "status is invalid");
    }

    private String buildSuccessResultJson(AICallbackRequest request) {
        try {
            return objectMapper.writeValueAsString(new SuccessPayload(
                    request.processingStats(),
                    request.cropRecommendations()
            ));
        } catch (Exception ex) {
            return "{\"status\":\"completed\"}";
        }
    }

    private String buildFailureResultJson(AICallbackRequest request) {
        try {
            return objectMapper.writeValueAsString(new FailedPayload(
                    request.errorCode(),
                    request.retryable()
            ));
        } catch (Exception ex) {
            return "{\"status\":\"failed\"}";
        }
    }

    private record SuccessPayload(Object processingStats, Object cropRecommendations) {
    }

    private record FailedPayload(String errorCode, Boolean retryable) {
    }
}